+++ main.py
@@ -21,9 +21,9 @@ from src.services import healthchecker
 
 @asynccontextmanager
 async def lifespan(fastapi_app: FastAPI):
-    # redis = aioredis.from_url(app_config.REDIS_URL, encoding='utf8')
-    # FastAPICache.init(RedisBackend(redis), prefix='fastapi-cache')
-    # await FastAPILimiter.init(redis)
+    redis = aioredis.from_url(app_config.REDIS_URL, encoding='utf8')
+    FastAPICache.init(RedisBackend(redis), prefix='fastapi-cache')
+    await FastAPILimiter.init(redis)
 
     yield
