+++ src/scores/models.py
@@ -5,16 +5,9 @@ from sqlalchemy.orm import Mapped, mapped_column
 from conf.config import Base
 
 
-
-# + id: int
-# + post_id
-# + user_id
-# + score
-
-
 class Score(Base):
     __tablename__ = "scores"
-    id: Mapped[int] = Column(Integer, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     post_id: Mapped[int] = mapped_column(Integer, ForeignKey("posts.id"), nullable=False)
     user_id: Mapped[int] = mapped_column(Integer, ForeignKey('users.id'), nullable=False)
     score: Mapped[int] = mapped_column(Integer, nullable=False)
